+++ components/RelatedWorksListItem/index.tsx
@@ -4,9 +4,10 @@ import { Button } from "react-aria-components";
 import styles from "./RelatedWorksListItem.module.scss";
 import DOMPurify from "dompurify";
 import ExpandButton from "@/components/ExpandButton";
-import { doiToUrl, orcidToUrl, rorToUrl } from "@/lib/idToUrl";
+import { doiToUrl, orcidToUrl, rorToUrl } from "@/lib/identifierUtils";
 import ExpandableNameList from "@/components/ExpandableNameList";
-import { Author, RelatedWorkSearchResult, RelatedWorkStatus, WorkVersion } from "@/generated/graphql";
+import { RelatedWorkSearchResult, RelatedWorkStatus } from "@/generated/graphql";
+import { formatAuthorNameFirstLast, formatSubtitle } from "@/lib/relatedWorks";
 
 const MAX_ITEMS = 10;
 const MAX_AUTHOR_CHARS = 40;
@@ -28,7 +29,9 @@ function RelatedWorksListItem({ relatedWork, highlightMatches, updateRelatedWork
   const expandedContentId = `${relatedWork.workVersion?.title?.toLowerCase().replace(/\s+/g, "-")}-content`;
   const headingId = `${relatedWork.workVersion?.title?.toLowerCase().replace(/\s+/g, "-")}-heading`;
   const { authorNames, containerTitle, publicationYear } = formatSubtitle(
-    relatedWork.workVersion as WorkVersion,
+    relatedWork.workVersion.authors,
+    relatedWork.workVersion.publicationVenue,
+    relatedWork.workVersion.publicationDate,
     MAX_AUTHOR_CHARS,
   );
 
@@ -50,7 +53,7 @@ function RelatedWorksListItem({ relatedWork, highlightMatches, updateRelatedWork
                 <h3 id={headingId}>
                   <a
                     href={doiToUrl(relatedWork.workVersion.work.doi)}
-                    aria-label={`${t("header.title")}: ${relatedWork.workVersion?.title}`}
+                    aria-label={`${t("title")}: ${relatedWork.workVersion?.title}`}
                     target="_blank"
                     rel="noopener noreferrer"
                   >
@@ -123,7 +126,7 @@ function RelatedWorksListItem({ relatedWork, highlightMatches, updateRelatedWork
               <Button
                 onPress={async () => {
                   setFadeOut(true);
-                  await updateRelatedWorkStatus(relatedWork.id, RelatedWorkStatus.Accepted);
+                  await updateRelatedWorkStatus(relatedWork.id as number, RelatedWorkStatus.Accepted);
                 }}
                 className={[relatedWork.status === RelatedWorkStatus.Pending ? "primary" : "secondary", "small"].join(
                   " ",
@@ -136,7 +139,7 @@ function RelatedWorksListItem({ relatedWork, highlightMatches, updateRelatedWork
               <Button
                 onPress={async () => {
                   setFadeOut(true);
-                  await updateRelatedWorkStatus(relatedWork.id, RelatedWorkStatus.Rejected);
+                  await updateRelatedWorkStatus(relatedWork.id as number, RelatedWorkStatus.Rejected);
                 }}
                 className={[relatedWork.status === RelatedWorkStatus.Pending ? "primary" : "secondary", "small"].join(
                   " ",
@@ -356,73 +359,6 @@ const useFormatDate = (date: Date | string | null | undefined) => {
 
   // Replace slashes with hyphens
   return formattedDate.replace(/\//g, "-");
-};
-
-const formatAuthorNameAbrev = (author: Author): string | null => {
-  const parts = [];
-
-  if (author.firstInitial && author.surname) {
-    // Combine initials
-    const initials = [];
-    if (author.firstInitial) {
-      initials.push(author.firstInitial);
-    }
-    if (author.middleInitials) {
-      initials.push(author.middleInitials);
-    }
-    parts.push(initials.join("")); // Join initials
-    parts.push(author.surname); // Add surname
-  } else if (author.full) {
-    // Fallback to full name
-    parts.push(author.full);
-  }
-
-  return parts.length > 0 ? parts.join(" ").trim() : null;
-};
-
-const formatAuthorNameFirstLast = (author: Author): string | null => {
-  const parts = [];
-
-  if (author.givenName && author.surname) {
-    parts.push(author.givenName);
-    parts.push(author.surname);
-  } else if (author.full) {
-    // Fallback to full name
-    parts.push(author.full);
-  }
-
-  return parts.length > 0 ? parts.join(" ").trim() : null;
-};
-
-const formatSubtitle = (
-  workVersion: WorkVersion,
-  maxAuthorChars: number,
-): { authorNames: string; containerTitle: string; publicationYear: string } => {
-  // Build author names
-  const names = workVersion.authors.map(formatAuthorNameAbrev).filter((n): n is string => !!n);
-
-  // Choose what names to display based on the total character length
-  const authorNames = [];
-  let totalChars = 0;
-  for (const name of names) {
-    authorNames.push(name);
-    totalChars += name.length;
-    if (totalChars >= maxAuthorChars) break;
-  }
-  if (names.length > authorNames.length) {
-    authorNames.push("et al");
-  }
-
-  // Build container title
-  let containerTitle = "";
-  if (workVersion.publicationVenue) {
-    containerTitle = ` ${workVersion.publicationVenue}${workVersion.publicationDate ? ", " : "."}`;
-  }
-
-  // Build publication year
-  const publicationYear = workVersion.publicationDate ? `${new Date(workVersion.publicationDate).getFullYear()}.` : "";
-
-  return { authorNames: authorNames.join(", ") + ". ", containerTitle, publicationYear };
 };
 
 export default RelatedWorksListItem;